requirements.txt
alembic>=1.14.0
asyncpg>=0.30.0
fastapi>=0.115.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.36
uvicorn>=0.32.0
